+++ backend/src/main/java/com/example/IOT_SmartStick/repository/DeviceRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface DeviceRepository extends JpaRepository<Device, Integer> {
+public interface DeviceRepository extends JpaRepository<Device, Long> {
     Optional<Device> findByDeviceToken(String deviceToken);
     List<Device> findByOwnerId(Integer ownerId);
     boolean existsByDeviceToken(String deviceToken);
